--- src/lib/components/icon/index.ts
@@ -1,2 +1,0 @@
-export { default as IconPicker } from './icon-picker.svelte';
-export * from './icon-picker.svelte';
